S4: promote ELF helpers from test code into ehmaps package
